requirements.txt
# Core
numpy
scipy
pandas
tqdm
pyyaml

# HuggingFace
datasets
transformers
accelerate
huggingface_hub

# RL / Preference learning
trl
peft
torch

# ML / stats
scikit-learn

# Visualization
matplotlib
seaborn
